automation-engine/src/ai/image_processor.py
"""
Image Processor
Handles image processing and classification tasks
"""
import logging
from typing import Dict, Any, List
import time
import cv2
import numpy as np
from PIL import Image

logger = logging.getLogger(__name__)


class ImageProcessor:
    """Handles image processing tasks"""
    
    def __init__(self):
        """Initialize image processor"""
        logger.info("Image Processor initialized")
    
    def classify_image(self, image_path: str) -> Dict[str, Any]:
        """
        Classify image (basic implementation using OpenCV)
        
        Args:
            image_path: Path to image file
            
        Returns:
            Dict containing classification results
        """
        start_time = time.time()
        
        try:
            # Load image
            img = cv2.imread(image_path)
            if img is None:
                raise ValueError(f"Could not load image: {image_path}")
            
            # Get basic image properties
            height, width, channels = img.shape
            
            # Calculate average color
            avg_color = img.mean(axis=0).mean(axis=0)
            
            # Determine dominant color
            dominant_color = self._get_dominant_color(avg_color)
            
            # Calculate brightness
            gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
            brightness = gray.mean()
            
            processing_time_ms = int((time.time() - start_time) * 1000)
            
            return {
                'width': int(width),
                'height': int(height),
                'channels': int(channels),
                'dominant_color': dominant_color,
                'brightness': float(brightness),
                'avg_color_bgr': [float(c) for c in avg_color],
                'processing_time_ms': processing_time_ms,
                'note': 'Basic classification - advanced ML models can be added'
            }
            
        except Exception as e:
            logger.error(f"Error in image classification: {e}")
            return {
                'error': str(e),
                'processing_time_ms': int((time.time() - start_time) * 1000)
            }
    
    def detect_objects(self, image_path: str) -> Dict[str, Any]:
        """
        Detect objects in image
        (Placeholder for future ML model integration)
        """
        return {
            'objects': [],
            'note': 'Object detection not yet implemented - can integrate YOLO or similar'
        }
    
    def enhance_image(self, image_path: str, output_path: str) -> Dict[str, Any]:
        """
        Enhance image quality
        
        Args:
            image_path: Input image path
            output_path: Output image path
            
        Returns:
            Dict containing enhancement results
        """
        start_time = time.time()
        
        try:
            # Load image
            img = cv2.imread(image_path)
            if img is None:
                raise ValueError(f"Could not load image: {image_path}")
            
            # Apply enhancement techniques
            # 1. Increase contrast
            lab = cv2.cvtColor(img, cv2.COLOR_BGR2LAB)
            l, a, b = cv2.split(lab)
            clahe = cv2.createCLAHE(clipLimit=3.0, tileGridSize=(8, 8))
            l = clahe.apply(l)
            enhanced = cv2.merge([l, a, b])
            enhanced = cv2.cvtColor(enhanced, cv2.COLOR_LAB2BGR)
            
            # 2. Denoise
            enhanced = cv2.fastNlMeansDenoisingColored(enhanced, None, 10, 10, 7, 21)
            
            # Save enhanced image
            cv2.imwrite(output_path, enhanced)
            
            processing_time_ms = int((time.time() - start_time) * 1000)
            
            return {
                'input_path': image_path,
                'output_path': output_path,
                'enhancements': ['contrast_adjustment', 'denoising'],
                'processing_time_ms': processing_time_ms
            }
            
        except Exception as e:
            logger.error(f"Error in image enhancement: {e}")
            return {
                'error': str(e),
                'processing_time_ms': int((time.time() - start_time) * 1000)
            }
    
    def _get_dominant_color(self, avg_color: np.ndarray) -> str:
        """Determine dominant color from BGR values"""
        b, g, r = avg_color
        
        if r > g and r > b:
            return 'red'
        elif g > r and g > b:
            return 'green'
        elif b > r and b > g:
            return 'blue'
        elif r > 200 and g > 200 and b > 200:
            return 'white'
        elif r < 50 and g < 50 and b < 50:
            return 'black'
        else:
            return 'mixed'
    
    def batch_process(self, image_paths: List[str], operation: str = 'classify') -> List[Dict[str, Any]]:
        """
        Process multiple images in batch
        
        Args:
            image_paths: List of image paths
            operation: Operation to perform ('classify', 'enhance')
            
        Returns:
            List of results for each image
        """
        results = []
        
        for image_path in image_paths:
            if operation == 'classify':
                result = self.classify_image(image_path)
            else:
                result = {'error': f'Unknown operation: {operation}'}
            
            results.append({
                'image_path': image_path,
                'result': result
            })
        
        return results
