automation-engine/requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0

# AI/ML dependencies
transformers==4.35.2
torch==2.1.1
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
sentencepiece==0.1.99

# Scripting language integration
lupa==2.0

# HTTP clients
requests==2.31.0
aiohttp==3.9.1

# Email
secure-smtplib==0.1.1

# Utilities
croniter==2.0.1
pytz==2023.3
pyyaml==6.0.1

# Logging
coloredlogs==15.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
